requirements.txt
python-telegram-bot>=20.0
supabase
pandas
openpyxl
google-generativeai
python-dotenv
